move build_filter/build_matcher to archiver.common
